urban-commute-assistant/backend/requirements.txt
Flask==2.1.1
fastapi==0.75.0
uvicorn==0.17.0
sqlalchemy==1.4.27
pydantic==1.9.0
httpx==0.21.1
python-dotenv==0.19.2
pytest==6.2.5
pytest-asyncio==0.15.1
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
requests==2.26.0
gtfs-realtime-bindings==1.0.0
pandas==1.4.2
redis==4.3.4
zipfile36==0.1.3
email-validator==1.1.3
python-multipart==0.0.5
python-dateutil==2.8.2
aiofiles==0.8.0
aiohttp==3.8.1
psycopg2-binary==2.9.9